simplify the app/models, duplicated paper recommendation
