+++ backend/app/routers/papers.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import List, Optional
 import re
 import logging
 from datetime import timedelta
@@ -9,8 +9,7 @@ from fastapi.responses import RedirectResponse
 from sqlalchemy.ext.asyncio import AsyncSession
 import httpx
 
-from ..models.users import User, UserPaperRecommendation
-from ..models.papers import PaperBase, PaperRecommendation
+from ..models.users import User, UserPaperRecommendation, PaperBase
 from ..db_utils import get_db, INDEX_SERVICE_URL
 from minio import Minio
 from minio.error import S3Error
@@ -19,10 +18,24 @@ import os
 from pydantic import BaseModel
 from datetime import datetime, timezone
 from fastapi.responses import RedirectResponse, Response
-from ..models.users import User, UserPaperRecommendation
-from ..models.papers import PaperBase, PaperRecommendation
+from ..models.users import User, UserPaperRecommendation, PaperBase
 from ..db_utils import get_db
 from ..auth.utils import get_current_user
+
+# Request model for adding recommendations
+class AddRecommendationRequest(BaseModel):
+    paper_id: str
+    title: Optional[str] = None
+    authors: Optional[str] = None
+    abstract: Optional[str] = None
+    url: Optional[str] = None
+    blog: Optional[str] = None
+    blog_abs: Optional[str] = None
+    blog_title: Optional[str] = None
+    recommendation_reason: Optional[str] = None
+    relevance_score: Optional[float] = None
+    submitted: Optional[str] = None
+    comment: Optional[str] = None
 
 # 设置日志
 logger = logging.getLogger(__name__)
@@ -254,7 +267,7 @@ async def serve_file(bucket: str, key: str):
 # 接口为{backend_url}/api/papers/recommend
 # TODO(@Hui Chen): 需要添加安全验证
 @router.post("/recommend", status_code=status.HTTP_201_CREATED)
-async def add_paper_recommendation(username:str, rec: PaperRecommendation, db: AsyncSession = Depends(get_db)):
+async def add_paper_recommendation(username:str, rec: AddRecommendationRequest, db: AsyncSession = Depends(get_db)):
     """根据username和paper详细信息插入推荐记录到UserPaperRecommendation表中"""
     try:
         # 验证用户是否存在
