+++ beta_frontend/nginx_mac.conf
@@ -12,7 +12,7 @@ http {
     }
 
     server {
-        listen 3000;
+        listen 8080;
         server_name localhost;
 
         # Static files root directory
@@ -65,6 +65,39 @@ http {
             proxy_send_timeout 60s;
             proxy_read_timeout 60s;
         }
+        # 文件服务代理
+        location /files/ {
+            proxy_pass http://backend;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            
+            # CORS headers for file requests
+            add_header Access-Control-Allow-Origin *;
+            add_header Access-Control-Allow-Methods "GET, OPTIONS";
+            add_header Access-Control-Allow-Headers "DNT,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Range";
+            
+            # Handle preflight requests for files
+            if ($request_method = OPTIONS) {
+                add_header Access-Control-Allow-Origin *;
+                add_header Access-Control-Allow-Methods "GET, OPTIONS";
+                add_header Access-Control-Allow-Headers "DNT,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Range";
+                add_header Access-Control-Max-Age 1728000;
+                add_header Content-Type 'text/plain; charset=utf-8';
+                add_header Content-Length 0;
+                return 204;
+            }
+            
+            # 文件缓存设置
+            proxy_cache_valid 200 1h;
+            proxy_cache_key "$scheme$request_method$host$request_uri";
+            
+            # 超时设置
+            proxy_connect_timeout 60s;
+            proxy_send_timeout 60s;
+            proxy_read_timeout 60s;
+        }
 
         # Handle WebSocket connections
         location /ws/ {
@@ -82,10 +115,11 @@ http {
         location / {
             try_files $uri $uri/ /index.html;
 
-            # Cache static resources
+            # No cache for development
             location ~* \.(js|css|png|jpg|jpeg|gif|ico|svg|woff|woff2|ttf|eot)$ {
-                expires 1y;
-                add_header Cache-Control "public, immutable";
+                expires -1;
+                add_header Cache-Control "no-cache, no-store, must-revalidate";
+                add_header Pragma "no-cache";
             }
         }
